requirements.txt
st-theme==1.2.3
streamlit~=1.40.0
streamlit-navigation-bar==3.2.0
pillow~=11.0.0